refactor(accelerator): extract bb download pipeline into versions/downloader.rs (F-04)
- versions.rs -> versions/mod.rs + versions/downloader.rs (download/verify/install pipeline)
- macOS xattr+codesign finalize tail extracted to finalize_downloaded_binary (no-op off macOS)
- pure move: download_bb re-exported; shared helpers pub(crate); consumers (bb/prove/tray) unchanged
- scope: 2-module split (heaviest concern) vs planned 5-way; identity/platform/layout/cache stay
  cohesive in mod.rs (phase-2 lessons)
- core 119 tests + clippy -D warnings green; server + src-tauri check clean
